requirements.txt
Flask>=3.0.0
Flask-SQLAlchemy>=3.1.1
Flask-Login>=0.6.3
Flask-Migrate>=4.0.7
psycopg2-binary>=2.9.9
PyMySQL>=1.1.1
python-dotenv>=1.0.1
qrcode>=7.4.2
Pillow>=10.4.0
